pyproject.toml
[tool.poetry]
name = "sxjm2024"
version = "0.1.0"
description = ""
authors = ["Wang Haowei <hw_wang@whu.edu.cn>"]
readme = "README.md"
packages = [{include = "traffic", from = "src"}]

[tool.poetry.dependencies]
python = "^3.8"
matplotlib = ">=3.2.2"
numpy = ">=1.18.5"
opencv-python = ">=4.1.2"
Pillow = ">=7.1.2"
PyYAML = ">=5.3.1"
requests = ">=2.23.0"
scipy = ">=1.4.1"
torch = ">=1.7.0"
torchvision = ">=0.8.1"
tqdm = ">=4.41.0"
tensorboard = ">=2.4.1"
pandas = ">=1.1.4"
seaborn = ">=0.11.0"
thop = "*"
easydict = "*"

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"

[tool.poetry.scripts]
detect = "cmd_line:main"

[tool.ruff]
line-length = 79